LayoutTests/fast/css3-text/css3-text-decoration/text-underline-position/underline-visual-overflow-with-subpixel-position-expected.txt

PASS window.internals.repaintRectsAsText().indexOf('8 8 784 12') is not -1
PASS window.internals.repaintRectsAsText().indexOf('8 20 784 12') is not -1
PASS successfullyParsed is true

TEST COMPLETE
foobar
foobar
